requirements.txt
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
easyocr==1.7.2
filelock==3.16.1
fsspec==2024.12.0
h11==0.14.0
idna==3.10
imageio==2.36.1
Jinja2==3.1.5
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.3
numpy==2.2.1
opencv-python-headless==4.10.0.84
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
pyclipper==1.3.0.post6
PySocks==1.7.1
python-bidi==0.6.3
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.25.0
scipy==1.15.1
selenium==4.27.1
shapely==2.0.6
sniffio==1.3.1
sortedcontainers==2.4.0
sympy==1.13.1
tifffile==2025.1.10
torch==2.5.1
torchvision==0.20.1
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
